[data] musan

+++ Define_Model/FilterLayer.py
@@ -567,7 +567,7 @@ class GaussianNoiseLayer(nn.Module):
             x_mean = torch.mean(x, dim=2, keepdim=True)
             x_std = torch.std(x, dim=2, keepdim=True)
 
-            gaussian_noise = torch.normal(mean=x_mean, std=x_std, dtype=torch.float32)
+            gaussian_noise = torch.normal(mean=x_mean, std=x_std)
             drop_weight = self.gaussion_weight.cuda() if x.is_cuda else self.gaussian_noise
             gaussian_noise *= drop_weight
